Moving uuid package to more accessible location.
